requirements.txt
 pytest>=8.0.0
 allure-pytest>=2.13.0
 selenium>=4.17.0
 webdriver-manager>=4.0.1
PyYAML>=6.0.1
